Add files via upload
This is todo application. in which user can add their daily tasks by category of personal or official. They can add task an an pending filter after complition they can edit it with an complete filter. also if the task is totally done or they need not to see that task they can delete it.